move tuning to adaptation
